ai-engine/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.27.0

# LangChain & LangGraph
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-core==0.1.10
langgraph==0.0.20

# ChromaDB
chromadb==0.4.22

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
